==> Step_2. finetuning_data.ipynb <==
import pandas as pd

fine_data1 = pd.read_excel('./data/학습모델2_연체.xlsx')
fine_data2 = pd.read_excel('./data/학습모델2_정상.xlsx')
fine_data = pd.concat([fine_data1,fine_data2], ignore_index=True)

fine_data

new_column_names = {
    'LIMIT_BAL': 'Amount of given credit in NT dollars',
    'SEX': 'Gender',
    'EDUCATION': 'Education',
    'MARRIAGE': 'Marital status',
    'AGE': 'Age',
    # 'PAY_0': 'Repayment status in September, 2005',
    # 'PAY_2': 'Repayment status in August, 2005',
    # 'PAY_3': 'Repayment status in July, 2005',
    'PAY_4': 'Repayment status in June, 2005',
    'PAY_5': 'Repayment status in May, 2005',
    'PAY_6': 'Repayment status in April, 2005',
    # 'BILL_AMT1': 'Amount of bill statement in September, 2005',
    # 'BILL_AMT2': 'Amount of bill statement in August, 2005',
    # 'BILL_AMT3': 'Amount of bill statement in July, 2005',
    'BILL_AMT4': 'Amount of bill statement in June, 2005',
    'BILL_AMT5': 'Amount of bill statement in May, 2005',
    'BILL_AMT6': 'Amount of bill statement in April, 2005',
    # 'PAY_AMT1': 'Amount of previous payment in September, 2005',
    # 'PAY_AMT2': 'Amount of previous payment in August, 2005',
    # 'PAY_AMT3': 'Amount of previous payment in July, 2005',
    'PAY_AMT4': 'Amount of previous payment in June, 2005',
    'PAY_AMT5': 'Amount of previous payment in May, 2005',
    'PAY_AMT6': 'Amount of previous payment in April, 2005',
    'default.payment.next.month': 'Default Payment Next Month'
}

fine_data = fine_data.rename(columns=new_column_names)

fine_data_input = fine_data.drop(columns=['Default Payment Next Month'])
fine_data_output = fine_data[['Default Payment Next Month']].replace({1: 'Yes, this customer will default on his/her credit card payment. ', 0: 'No, this  customer will not default on his/her credit card payment.'})

def tab_to_str(row):
    output = ""
    for col, val in row.items():
        output += f"{col}: {val}\n"
    return output[:-1]


tab_to_str_input_list = []
for index, row in fine_data_input.iterrows():
    tab_to_str_input_list.append(tab_to_str(row))

tab_to_str_input_list_df = pd.DataFrame(tab_to_str_input_list, columns=["list"])

fine_data_output = fine_data_output.reset_index(drop=True)

final_data = pd.concat([tab_to_str_input_list_df, fine_data_output], axis=1)

final_data["Default Payment Next Month"][0]

final_data

print(final_data["list"][0])

final_data["Default Payment Next Month"]

# system_prompt = "You are a helpful AI assistant for financial and  commercial analysis"

# user_prompt = f"""when the customer's conditions are as follows:
# {final_data["list"]}
# Will the customer be late on October 2005 payment?"""

# base_prompt = f"<|begin_of_text|><|start_header_id|>system<|end_header_id|>{system_prompt}<|eot_id|><|start_header_id|>user<|end_header_id|>{user_prompt}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"

def generate_prompt(row):
    input = row["list"]
    output = row["Default Payment Next Month"]

    system_prompt = "You are a helpful AI assistance for financial and commercial analysis"
    user_prompt = f"when the customer's conditions are as follows:\n{input}\nWill the customer be late on October 2005 payment?"



    base_prompt = f"<|begin_of_text|><|start_header_id|>system<|end_header_id|>\n{system_prompt}<|eot_id|>\n<|start_header_id|>user<|end_header_id|>\n{user_prompt}<|eot_id|><|start_header_id|>assistant<|end_header_id|>\n{output}<|eot_id|>"
    
    return base_prompt

final_data

# Apply the function to each row in the DataFrame
final_data["prompt"] = final_data.apply(generate_prompt, axis=1)

print(final_data["prompt"][0])

final_data.to_pickle('./data/finetuning_data_2.pkl')
