==> Step_3. fine tuning_code.ipynb <==
https://www.kaggle.com/code/ssarkar445/huggingface-tinyllama-finetune-peft

!pip install -q bitsandbytes accelerate loralib trl 
!pip install -q git+https://github.com/huggingface/peft.git
!pip install -U git+https://github.com/huggingface/transformers.git
!pip install -U git+https://github.com/huggingface/accelerate.git
!pip install -U datasets

import os
import torch
from datasets import Dataset
from peft import LoraConfig,AutoPeftModelForCausalLM,PeftModel
from transformers import AutoModelForCausalLM,AutoTokenizer,BitsAndBytesConfig,TrainingArguments,pipeline
from trl import SFTTrainer
from time import perf_counter
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
device

import gc
gc.collect()
torch.cuda.empty_cache()

import os
class myconfig():
    data_path = r"/home/work/lib_data/hjy/finetuning_data_v4.pkl"
    access_token = os.getenv('HUGGINGFACE_TOKEN')
    model = "meta-llama/Meta-Llama-3-8B-Instruct"
    final_model = "KIML/llama3_8B_hjy_v2"
    cache_dir = r"/home/work/lib_data/Llama3"
    save_path = r"/home/work/lib_data/hjy/finetuned_model"
    result_path_not_late = r"/home/work/lib_data/hjy/results_not_late.pkl"
    result_path_late = r"/home/work/lib_data/hjy/results_late.pkl"
    
    result_path_not_late_2 = r"/home/work/lib_data/hjy/results_not_late_2.pkl"
    result_path_late_2 = r"/home/work/lib_data/hjy/results_late_2.pkl"
    seed = 117
mcfg = myconfig()

import pandas as pd
finetuning_data_v3 = pd.read_pickle(mcfg.data_path)

print(finetuning_data_v3["prompt"][0])

from sklearn.model_selection import train_test_split
train_df, eval_df = train_test_split(
    finetuning_data_v3,
    test_size=0.2,
    random_state=mcfg.seed,
    stratify=finetuning_data_v3['Default Payment Next Month']
)

train_df

train_dataset = Dataset.from_pandas(pd.DataFrame(train_df["prompt"]))
eval_dataset = Dataset.from_pandas(pd.DataFrame(eval_df["prompt"]))

import torch
torch.cuda.is_available()

from huggingface_hub import login
login(mcfg.access_token)

# 학습

def get_tokenizer_and_model(model_id):
    tokenizer = AutoTokenizer.from_pretrained(model_id, use_auth_token=True, cache_dir=mcfg.cache_dir)
    tokenizer.pad_token = tokenizer.eos_token
    bnb_config = BitsAndBytesConfig(
        load_in_4bit=True, bnb_4bit_quant_type="nf4", bnb_4bit_compute_dtype="float16", bnb_4bit_use_double_quant=True
    )
    model = AutoModelForCausalLM.from_pretrained(model_id,quantization_config=bnb_config, device_map="auto", cache_dir=mcfg.cache_dir)
    model.config.use_cache=False
    model.config.pretraining_tp=1
    return model,tokenizer

model, tokenizer = get_tokenizer_and_model(mcfg.model)

peft_config = LoraConfig(
    r=8,lora_alpha=16,lora_dropout=0.05,bias="none",task_type="CAUSAL_LM"
)

training_arguments = TrainingArguments(
    output_dir=mcfg.save_path,
    per_device_train_batch_size=8,
    gradient_accumulation_steps=4,
    optim="paged_adamw_32bit",
    learning_rate=2e-4,
    lr_scheduler_type="cosine",
    evaluation_strategy="epoch",
    logging_strategy="epoch",
    save_strategy="epoch",
    logging_steps=10,
    num_train_epochs=50,
    # max_steps=200,
    fp16=True,
    push_to_hub=False
)

trainer = SFTTrainer(
        model=model,
        train_dataset=train_dataset,
        eval_dataset=eval_dataset,
        peft_config=peft_config,
        dataset_text_field="prompt",
        args=training_arguments,
        tokenizer=tokenizer,
        packing=False,
        max_seq_length=1024
    )

trainer.train()

# trainer.evaluate()

model = AutoModelForCausalLM.from_pretrained(mcfg.model,torch_dtype=torch.float16, load_in_8bit=False,
                                             device_map="auto",
                                             trust_remote_code=True,
                                             cache_dir=mcfg.cache_dir)

peft_model = PeftModel.from_pretrained(model,
                                       '/home/work/lib_data/hjy/finetuned_model/checkpoint-275',
                                       from_transformers=True,
                                       device_map={"":0}
                                       )

model = peft_model.merge_and_unload()

model.push_to_hub(mcfg.final_model,token=mcfg.access_token)

tokenizer = AutoTokenizer.from_pretrained(mcfg.model)
tokenizer.push_to_hub(mcfg.final_model,token=mcfg.access_token)
