==> Step_4. inference.ipynb <==
from numba import cuda
reset_d = cuda.get_current_device()
reset_d.reset()

import os
class myconfig():
    data_path = r"/home/work/lib_data/KISTI/summary_data_ctgr.pkl"
    access_token = os.getenv('HUGGINGFACE_TOKEN')
    model = "KIML/llama3_8B_hjy_v2"
    cache_dir = r"/home/work/lib_data/Llama3"
mcfg = myconfig()

import pandas as pd
summary_data = pd.read_pickle(mcfg.data_path)
title = list(summary_data["title"])
summary = list(summary_data["summary"])

import torch
torch.cuda.empty_cache()
torch.cuda.is_available()

from huggingface_hub import login
login(mcfg.access_token)

# 1. 모델 불러오기

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

import transformers
from transformers import AutoTokenizer, AutoModelForCausalLM
import numpy as np

tokenizer = AutoTokenizer.from_pretrained(mcfg.model, cache_dir=mcfg.cache_dir)
model = AutoModelForCausalLM.from_pretrained(mcfg.model, cache_dir=mcfg.cache_dir)

model.to(device)

# 2. 프롬프트 생성

system_prompt = """You are a helpful AI assistant for financial and commercial analysis."""
user_prompt = f"""If customer does not make late payments in October 2005, generate the values ​​of the following conditions:

Amount of given credit in NT dollars: 
Gender: 
Education: 
Marital status: 
Age: 
Repayment status in June, 2005: 
Repayment status in May, 2005: 
Repayment status in April, 2005: 
Amount of bill statement in June, 2005: 
Amount of bill statement in May, 2005: 
Amount of bill statement in April, 2005: 
Amount of previous payment in June, 2005: 
Amount of previous payment in May, 2005: 
Amount of previous payment in April, 2005: 

Do not create repeated content or multiple conditions.
"""
    
base_prompt = f"<|begin_of_text|><|start_header_id|>system<|end_header_id|>{system_prompt}<|eot_id|><|start_header_id|>user<|end_header_id|>{user_prompt}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"

input = tokenizer(base_prompt, return_tensors="pt").to(device)

# input = tokenizer(user_prompt, return_tensors="pt").to(device)

# 3. 추론(generation)

# from transformers import pipeline

# pipe = pipeline(
#     "text-generation",
#     model=model,
#     tokenizer=tokenizer,
#     torch_dtype=torch.float16,
#     device=device,
#     )

import gc
gc.collect()
torch.cuda.empty_cache()

results = []

for i in range(2000):
    generated_outputs = model.generate(**input,
                                    max_new_tokens = 200,
                                # return_dict_in_generate=True
    )

    # generated_outputs.sequences = input+output
    input_length = 1 if model.config.is_encoder_decoder else input.input_ids.shape[1]

    generated_tokens = generated_outputs[:,input_length:]

    sentence = tokenizer.decode(generated_tokens[0], skip_special_tokens=True)
    
    results.append(sentence)

    gc.collect()
    torch.cuda.empty_cache()

len(results)

result_no_late = pd.DataFrame(results)

result_no_late.to_csv("./result_no_late.csv")

results_2 = []

system_prompt = """You are a helpful AI assistant for financial and commercial analysis."""
user_prompt = f"""If customer will default on his/her credit card payment in October 2005, fill the [] with values ​​of the following conditions:

Amount of given credit in NT dollars: []
Gender: []
Education: [] 
Marital status: []
Age: []
Repayment status in June, 2005: []
Repayment status in May, 2005: []
Repayment status in April, 2005: []
Amount of bill statement in June, 2005: []
Amount of bill statement in May, 2005: []
Amount of bill statement in April, 2005: []
Amount of previous payment in June, 2005: []
Amount of previous payment in May, 2005: []
Amount of previous payment in April, 2005: []

Do not generate repeated content or multiple conditions.
"""
    
base_prompt = f"<|begin_of_text|><|start_header_id|>system<|end_header_id|>{system_prompt}<|eot_id|><|start_header_id|>user<|end_header_id|>{user_prompt}<|eot_id|><|start_header_id|>assistant<|end_header_id|>"

input = tokenizer(base_prompt, return_tensors="pt").to(device)

# input = tokenizer(user_prompt, return_tensors="pt").to(device)

for i in range(2000):
    generated_outputs = model.generate(**input,
                                    max_new_tokens = 250,
                                # return_dict_in_generate=True
    )

    # generated_outputs.sequences = input+output
    input_length = 1 if model.config.is_encoder_decoder else input.input_ids.shape[1]

    generated_tokens = generated_outputs[:,input_length:]

    sentence = tokenizer.decode(generated_tokens[0], skip_special_tokens=True)
    
    print(sentence)
    
    results_2.append(sentence)

    gc.collect()
    torch.cuda.empty_cache()

len(results_2)

result_late = pd.DataFrame(results_2)
result_late.to_csv("./result_late.csv")
