==> Step_5. inference_post_processing.ipynb <==
import pandas as pd

class myconfig():
    no_late_path = r"/home/work/lib_data/hjy/result_no_late.csv"
    no_late_save_path = "./no_late_df.xlsx"
    late_path = r"/home/work/lib_data/hjy/result_late.csv"
    late_save_path = "./late_df.xlsx"
mcfg = myconfig()

no_late_df = list(pd.read_csv(mcfg.no_late_path)["0"])
late_df = list(pd.read_csv(mcfg.late_path)["0"])

print(len(no_late_df))
print(len(late_df))

late_df

import re
import pandas as pd

df_no_late = pd.DataFrame()


for text in no_late_df:
    try:
        data = {
            "Amount of given credit in NT dollars": [int(re.search(r'Amount of given credit in NT dollars: (\d+)', text).group(1))],
            "Gender": [int(re.search(r'Gender: (\d+)', text).group(1))],
            "Education": [int(re.search(r'Education: (\d+)', text).group(1))],
            "Marital status": [int(re.search(r'Marital status: (\d+)', text).group(1))],
            "Age": [int(re.search(r'Age: (\d+)', text).group(1))],
            "Repayment status in October 2005": [int(re.search(r'Repayment status in October 2005: (\d+)', text).group(1))],
            "Repayment status in September 2005": [int(re.search(r'Repayment status in September 2005: (\d+)', text).group(1))],
            "Repayment status in August 2005": [int(re.search(r'Repayment status in August 2005: (\d+)', text).group(1))],
            "Amount of bill statement in October 2005": [int(re.search(r'Amount of bill statement in October 2005: (\d+)', text).group(1))],
            "Amount of bill statement in September 2005": [int(re.search(r'Amount of bill statement in September 2005: (\d+)', text).group(1))],
            "Amount of bill statement in August 2005": [int(re.search(r'Amount of bill statement in August 2005: (\d+)', text).group(1))],
            "Amount of previous payment in October 2005": [int(re.search(r'Amount of previous payment in October 2005: (\d+)', text).group(1))],
            "Amount of previous payment in September 2005": [int(re.search(r'Amount of previous payment in September 2005: (\d+)', text).group(1))],
            "Amount of previous payment in August 2005": [int(re.search(r'Amount of previous payment in August 2005: (\d+)', text).group(1))]
        }
    # Convert the dictionary into a DataFrame
        temp_df = pd.DataFrame(data)
    
    # Append the new row to the main DataFrame
        df_no_late = pd.concat([df_no_late, temp_df], ignore_index=True)    
    except:
        pass

df_no_late = df_no_late.sample(1000, random_state=42)    
    
df_no_late.to_excel(mcfg.no_late_save_path, index=False)

import re
import pandas as pd

df_late = pd.DataFrame()


for text in late_df:
    try:
        data = {
            "Amount of given credit in NT dollars": [int(re.search(r'Amount of given credit in NT dollars: (\d+)', text).group(1))],
            "Gender": [int(re.search(r'Gender: (\d+)', text).group(1))],
            "Education": [int(re.search(r'Education: (\d+)', text).group(1))],
            "Marital status": [int(re.search(r'Marital status: (\d+)', text).group(1))],
            "Age": [int(re.search(r'Age: (\d+)', text).group(1))],
            "Repayment status in October 2005": [int(re.search(r'Repayment status in October 2005: (\d+)', text).group(1))],
            "Repayment status in September 2005": [int(re.search(r'Repayment status in September 2005: (\d+)', text).group(1))],
            "Repayment status in August 2005": [int(re.search(r'Repayment status in August 2005: (\d+)', text).group(1))],
            "Amount of bill statement in October 2005": [int(re.search(r'Amount of bill statement in October 2005: (\d+)', text).group(1))],
            "Amount of bill statement in September 2005": [int(re.search(r'Amount of bill statement in September 2005: (\d+)', text).group(1))],
            "Amount of bill statement in August 2005": [int(re.search(r'Amount of bill statement in August 2005: (\d+)', text).group(1))],
            "Amount of previous payment in October 2005": [int(re.search(r'Amount of previous payment in October 2005: (\d+)', text).group(1))],
            "Amount of previous payment in September 2005": [int(re.search(r'Amount of previous payment in September 2005: (\d+)', text).group(1))],
            "Amount of previous payment in August 2005": [int(re.search(r'Amount of previous payment in August 2005: (\d+)', text).group(1))]
        }
    # Convert the dictionary into a DataFrame
        temp_df = pd.DataFrame(data)
    
    # Append the new row to the main DataFrame
        df_late = pd.concat([df_late, temp_df], ignore_index=True)    
    except:
        pass

df_late = df_late.sample(1000, random_state=42)  
    
df_late.to_excel(mcfg.late_save_path, index=False)

df_late
